LDLib2: NeoForge only. Consider matrix transforming.

+++ common/src/main/java/io/github/reserveword/imblocker/common/gui/Point.java
@@ -3,7 +3,7 @@ package io.github.reserveword.imblocker.common.gui;
 public record Point(int x, int y) {
 	public static final Point TOP_LEFT = new Point(0, 0);
 	
-	public Point(double scaleFactor, int x, int y) {
+	public Point(double scaleFactor, double x, double y) {
 		this((int) (scaleFactor * x), (int) (scaleFactor * y));
 	}
